New Development: no             JIRA Issue: CAS-10031       Ready to Release: yes      Interface Changes: none What Interface Changed: none          Test Programs: none   Put in Release Notes: no
           Description: add a delay between sigterm and sigkill, add debugging info when sigkill is needed



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@39968 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/procmgr.py
@@ -47,9 +47,12 @@ class procmgr(Thread):
                     #print "%s => proc %s is being stopped" % (strftime("%y-%m-%d %H:%M:%S", localtime()), self.tag)
                     try:
                         self.__proc.terminate()
-                        self.__watchdog.terminate()
-                        self.__proc.kill()
-                        self.__watchdog.kill()
+                        sleep(0.5)
+                        if self.__proc.poll( ) is None:
+                            print "%s => proc %s is being killed" % (strftime("%y-%m-%d %H:%M:%S", localtime()), self.tag)
+                            self.__proc.kill()
+                        if self.__watchdog.poll( ) is None:
+                            self.__watchdog.kill()
                     except OSError:
                         pass
